requirements.txt
pathlib==1.0.1
Pillow>=10.0.0  # 用于图像处理 